Add test for comments counter
