+++ package.json
@@ -33,7 +33,7 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.28.0",
-    "hint": "^7.1.10",
+    "hint": "^2.0.0",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.6.2",
     "jest-environment-jsdom": "^29.6.2",
