+++ src/modules/loadCharacters.js
@@ -142,33 +142,6 @@ const countCharacters = async () => {
   characterCountElement.textContent = `Characters(${characters.length})`;
 };
 
-const countComments = async (data) => {
-  const commentCountElement = document.getElementById('comment-counter');
-  commentCountElement.textContent = `Comments(${data.length < 1 ? 0 : data.length})`;
-};
-
-export const fetchComments = async (appId, characterId) => {
-  try {
-    const queryString = `?item_id=${characterId}`;
-    const url = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}/comments${queryString}`;
-
-    const response = await fetch(url);
-
-    if (response.ok) {
-      const data = await response.json();
-      countComments(data);
-      return data.map((comment) => {
-        const formattedDate = new Date(comment.creation_date).toLocaleDateString();
-        return `${formattedDate} ${comment.username}: ${comment.comment}`;
-      });
-    }
-    const errorMessage = await response.text();
-    throw new Error(`Failed to fetch comments: ${errorMessage}`);
-  } catch (error) {
-    return [];
-  }
-};
-
 export const updateModalContent = async (character) => {
   const modalBody = document.querySelector('.modal-body');
   if (!modalBody) {
